[Process] Split pipes management into UnixPipes and WindowsPipes
